- Clean up .gitignore to ignore CCS and user specific files. - Remove previously (but incorrectly) tracked files from the repository. - Standardise src folder structure for all teams. - Move non-satellite code out of this repository. - Create new scheme for defining headers. - Create private and public headers for all files, to assist in code modularity and encapsulation.
Create basic implementation of dynamic task allocator. Prototype stage, and still has issues to resolve such as the (limited) use of function pointers.

Create a rudimentary logging framework

Experiment with creating tasks dynamically, based somewhat on the
structure we discussed with Craig over the weekend.

In this commit, we set up the CDH system to be able to request tasks
whenever the master task allocator comes around asking. When the task
allocator asks CDH if it would like to submit a task, the CDH provided
interface returns a struct containing some information about the task it
would like to run.

Upon receiving this, the master allocator creates the task if there is
room available, and returns the task handle back to CDH. This allows CDH
to manage the task, such as killing it, or changing its priority. In
this example, CDH kills the task after a single run. In the future, we
may need to restrict the usage of the task priority change
functionality, or at least set guidelines on how it should be used.

Start working on UART.