Use the default Debug build configuration in CCS, instead of the custom name we gave it (TIRTOS Build).
This will help increase ease of use, as programmers no longer need to
manually change the build profile when configuring a project, and will
no longer receive cryptic errors if they forget to do so.

Keep the old TIRTOS Build configuration to maintain compatability with
those who are already on it, as we do not want to break their builds.

This fixes CS1-208.