Implement thread safe debug message queue
Enqueuing does not enqueue the actual message, but instead serialises and
then enqueues.

The purpose of serialising to heap before enqueuing is to avoid an
otherwise necessary deep copy of the abstract message pointer passed in.

A deep copy would be required to prevent the queue from being corrupted
if a message passed in were to go out of scope

Furthermore, a deep copy requires changing the message class such that all
derived classes implement a clone function.

The actual serialisation to heap involves separate manual allocation of
1) The serialised message buffer
2) The rest of the serialised message object

In addition, the queue itself is constructed on the heap to be persistent.

As such, there is a bit of dirty memory management, where each part allocated
in the EnqueueSerialised function should be correspondingly destroyed in
the Dequeue function.

Fortunately, the actual serialised message buffer is never copied as the
message serialises directly to it on the heap. However, any extra data
(which is not much) in the SerialisedMessage object does get cloned from
stack to heap.

Potential multi-thread initialisation error is solved by initialising the DebugStream singleton instance in the post bios init