Update the cppcheck configuration
This commit disables the `unusedFunction` warning as it is too noisy,
and while it does find technical violations their are operational
reasons the functions are left unused. This means that the output from
this check is effectively ignored. There is also no significant benefit
to having it enabled; it makes code easier to read for us, but not
difference to the compiler.

This commit also sets the `--force` flag on cppcheck which will make it
check all header files, potentially detecting more errors. This will
slow the analysis down somewhat but this is an acceptable trade-off, as
it currently takes only a few seconds to run.