Add rebuildable message functionality. CS1-423
This commit implements the inverse of the logic provided by
SerialisedMessage. All messages are serialisable but not all messages
are rebuildable so we split this logic into two separate classes,
related by inheritance (implementation <- RebuildableMessage <- Message,
or implementation <- Message).

Some things to watch out for:
- a constructor to rebuild a message is provided, but it must call
  `SerialiseFrom`. This needs to be added manually, as it was not
possible to use language features to enforce it.
- the ordering of rebuild must be the exact inverse of serialise.
- the data types must be exactly the same.