Configure MSP432 to use HW accessible I2C modules
This change will configure the I2C pin definitions to match those used
on our hardware. It also gives the buses names that are consistent with
the naming on the system wiring diagram.

This change also maintains backwards compatability by redirecting the
currently used bus to one that is also valid on the LaunchPad and PCBA.