Convert #define to feature flags for unit tests. CS1-733
The use of `bool` allows for a robust enable/disable of satellite features. `#define` is error prone and has unsafe failure conditions (eg. if the `unit_tests.h` file is mistakenly not `#included` then it will default to undefined, which actually enables our tests when the safe thing would be to disable them). Forgetting, misspelling, or any other problems will now be captured by a compilation warning instead.