Create I2C bus manager. CS1-769.
The bus manager will fix bugs that we see in day-to-day testing. It will
check if the I2C bus is physically present before attempting to use it.
Not doing this often causes the LaunchPad to hang as it attempts to make
I2C transactions on a bus that isn't there. The underlying I2C timeout
does not work here either, so we need to ensure the I2C transaction is
never attempted in the first place.

This also adds support for multiple threads to use the bus at once. Each
transaction is atomic but outside of that no protection is given. If
there is a need to support larger atomic transfers then we will need to
add semaphore functionality.