State determined by both BMS temperatures
State changes are determined by both BMSs. When the temperature is rising, the hottest BMS determines the state. When cooling, both BMSs have to be below the threshold. 

To do this it is necessary to identify the sensor returned from GetSensorWithReading(). The state machines stores the temperature from both BMSs and uses both temperatures to update its state.