Introduce failure mode for sensors and measurables
Currently no way to exit failure mode, likely to be implemented in the future if needed

`I2cSensor` class now contains a `failed` flag which can be set in the constructor on a failed initialisation (e.g. when ADC fails to set its config in constructor)

Downstream measurables which register themselves with sensors will check for the state of this flag and then set their own `failed` flag in accordance with this.

`failed==true` measurables will return a failure constant without attempting to read from hardware.

This allows all sensors to initialise, fail, and still have the rest of the system proceed as normal.

Unfortunately it does not protect against any flight software logic which utilizes readings for decision making, failure constants are typically way outside normal bounds (e.g. -9999) so it is up to the decision making code to perform sanity checks.

Do not check sensor failed for mocked sensors