Implement Science Data Command
In order to find out the correct nanopb type, ScienceDataCommand iterates through all possible downlink nanopb types and uses the first one that can be successfully read back from a circular buffer and decoded.

This may not necessarily be the 'same' nanopb in name as the one that is stored, as nanopb messages with the same structure can be decoded/encoded to each other.

However it is guaranteed to be able to be decoded on the ground using its original nanopb message type despite potentially being 'mistyped' temporarily.