Implement sectionwise over-the-air uplink
*Changelist/Notable:*

Add payload section manager
Increase payload processing stack size to avoid overflow
Usage of recursion in executing section commands
Unit tests which cover multiple cases

*Context*

We would like to be able to build a Command by breaking up the Command payload into smaller sections and uplinking the smaller sections.
My understanding is that smaller payloads sent from the ground station are more likely to be received by the Lithium.

Sections will be sent from the ground stations using SectionUplink. Received sections will be added to a buffer so that the 'Command' payload can be recreated.

*Problem*

Since we will be unsure which uplinks will be received by the Lithium we could receive random sections of the Command.
The manager needs to be able to store random sections of the payload in the correct order.

Once the the Command has been recreated it will be executed using ExecuteSectionsUplink

*How this will work at a high level:*

At the ground station we will break a single payload into multiple sections.
The ground station will uplink a SectionUplink which will contain 1 section.
This will be repeated for each section.
Note: It may be that not every uplink command will reach the satellite. It may be necessary to send each command multiple times.
When we think every section is uploaded we can query the section buffer and see if the entire payload has been recreated on the satellite. This will be done by QuerySectionsUplink
Note: Sending big payloads is not an issue on the downlink.
When we confirm that the payload has been recreated properly we can directly call it using the payload processor (to avoid HMAC and FEC checks).
We can clear the buffer using ClearSectionUplink.

*TODOs*

Check the buffer for the correct length corresponding to the command
code before execution
Ensure iterator does not result in an infinite loop

Co-authored-by: Will Schuetz <willschuetz@outlook.com>
Co-authored-by: Benjamin Ding <ben@dingbenjamin.com>