Dynamically change battery charge current
Enables potentially much more power to be harvested in the space
environment and reduces the risk in choosing a fixed charging rate.

Although varying the charging current does induce a 10-20% reduction in
charging time due to the interruption when a change is actioned,
being able to charge at higher rates when more power is available
offsets this.